requirements.txt
fastapi
uvicorn[standard]
whisper
llama-cpp-python
soundfile